@@ -5,17 +5,19 @@ namespace Database\Factories;
 use App\Models\Team;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
 use Laravel\Jetstream\Features;
 
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\User>
+ */
 class UserFactory extends Factory
 {
     /**
-     * The name of the factory's corresponding model.
-     *
-     * @var string
+     * The current password being used by the factory.
      */
-    protected $model = User::class;
+    protected static ?string $password;
 
     /**
      * Define the model's default state.
@@ -25,11 +27,10 @@ class UserFactory extends Factory
     public function definition(): array
     {
         return [
-            'name' => $this->faker->name(),
-            'email' => $this->faker->unique()->safeEmail(),
+            'name' => fake()->name(),
+            'email' => fake()->unique()->safeEmail(),
             'email_verified_at' => now(),
-            'last_activity' => now(),
-            'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
+            'password' => static::$password ??= Hash::make('password'),
             'two_factor_secret' => null,
             'two_factor_recovery_codes' => null,
             'remember_token' => Str::random(10),
@@ -39,51 +40,13 @@ class UserFactory extends Factory
     }
 
     /**
-     * Indicate that the user should be a admin.
-     */
-    public function admin(): static
-    {
-        return $this->state(function () {
-            return [
-                'admin' => now(),
-            ];
-        });
-    }
-
-    /**
-     * Indicate that the user should be a super admin.
-     */
-    public function superAdmin(): static
-    {
-        return $this->state(function () {
-            return [
-                'super_admin' => now(),
-            ];
-        });
-    }
-
-    /**
-     * Indicate that the user should be a admin.
-     */
-    public function dmin(): static
-    {
-        return $this->state(function () {
-            return [
-                'admin' => now(),
-            ];
-        });
-    }
-
-    /**
      * Indicate that the model's email address should be unverified.
      */
     public function unverified(): static
     {
-        return $this->state(function (array $attributes) {
-            return [
-                'email_verified_at' => null,
-            ];
-        });
+        return $this->state(fn (array $attributes) => [
+            'email_verified_at' => null,
+        ]);
     }
 
     /**
