@@ -2,22 +2,17 @@
 
 namespace App\Actions\Fortify;
 
-use Illuminate\Support\Facades\App;
-use Laravel\Fortify\Rules\Password;
+use Illuminate\Validation\Rules\Password;
 
 trait PasswordValidationRules
 {
     /**
      * Get the validation rules used to validate passwords.
      *
-     * @return array<int, \Illuminate\Contracts\Validation\Rule|array|string>
+     * @return array<int, \Illuminate\Contracts\Validation\Rule|array<mixed>|string>
      */
-    protected function passwordRules()
+    protected function passwordRules(): array
     {
-        if (! App::environment('local')) {
-            return (new Password())->length(10)->requireUppercase()->requireNumeric()->requireSpecialCharacter();
-        }
-
-        return ['required', 'string', new Password(), 'confirmed'];
+        return ['required', 'string', Password::default(), 'confirmed'];
     }
 }
