@@ -33,7 +33,7 @@ class RemoveTeamMember implements RemovesTeamMembers
     {
         if (! Gate::forUser($user)->check('removeTeamMember', $team) &&
             $user->id !== $teamMember->id) {
-            throw new AuthorizationException();
+            throw new AuthorizationException;
         }
     }
 
