@@ -2,10 +2,8 @@
 
 namespace App\Models;
 
-use App\Casts\Money;
-use Illuminate\Contracts\Auth\MustVerifyEmail;
+// use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Fortify\TwoFactorAuthenticatable;
@@ -13,7 +11,7 @@ use Laravel\Jetstream\HasProfilePhoto;
 use Laravel\Jetstream\HasTeams;
 use Laravel\Sanctum\HasApiTokens;
 
-class User extends Authenticatable implements MustVerifyEmail
+class User extends Authenticatable
 {
     use HasApiTokens;
     use HasFactory;
@@ -31,7 +29,6 @@ class User extends Authenticatable implements MustVerifyEmail
         'name',
         'email',
         'password',
-        'total_income_amount',
     ];
 
     /**
@@ -47,16 +44,6 @@ class User extends Authenticatable implements MustVerifyEmail
     ];
 
     /**
-     * The attributes that should be cast.
-     *
-     * @var array<string, string>
-     */
-    protected $casts = [
-        'email_verified_at' => 'datetime',
-        'total_income_amount' => Money::class,
-    ];
-
-    /**
      * The accessors to append to the model's array form.
      *
      * @var array<int, string>
@@ -65,45 +52,16 @@ class User extends Authenticatable implements MustVerifyEmail
         'profile_photo_url',
     ];
 
-    public function isAdmin()
-    {
-        return $this->admin != null ? true : false;
-    }
-
-    public function isSuperAdmin()
-    {
-        return $this->super_admin != null ? true : false;
-    }
-
     /**
-     * Gets the users their piggy banks
+     * Get the attributes that should be cast.
+     *
+     * @return array<string, string>
      */
-    public function piggyBanks(): HasMany
+    protected function casts(): array
     {
-        return $this->hasMany(PiggyBank::class);
-    }
-
-    /**
-     * Gets the users their income transactions
-     */
-    public function income(): HasMany
-    {
-        return $this->hasMany(Income::class);
-    }
-
-    /**
-     * Gets the users their expense transactions
-     */
-    public function category(): HasMany
-    {
-        return $this->hasMany(Category::class);
-    }
-
-    /**
-     * Gets the all the transactions that the user has over all categories
-     */
-    public function getTransactions()
-    {
-        return Transaction::whereIn('category_id', $this->category->pluck('id'))->get();
+        return [
+            'email_verified_at' => 'datetime',
+            'password' => 'hashed',
+        ];
     }
 }
