@@ -16,7 +16,7 @@ return [
     |
     */
 
-    'stack' => 'inertia',
+    'stack' => 'livewire',
 
     /*
      |--------------------------------------------------------------------------
@@ -59,7 +59,7 @@ return [
 
     'features' => [
         // Features::termsAndPrivacyPolicy(),
-        Features::profilePhotos(),
+        // Features::profilePhotos(),
         // Features::api(),
         Features::teams(['invitations' => true]),
         Features::accountDeletion(),
